backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pydantic==2.9.2
pydantic-settings==2.6.1
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# better-auth==0.2.0
better-auth==0.0.1b11
alembic==1.13.3
psycopg2-binary==2.9.10
python-multipart==0.0.20
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
slowapi==0.1.9